[refs #40] Added section breaks and width to core.
We had no style for <hr>'s so I created a Sass partial in core named _section-break.
There are varying sizes which use our spacing @includes to keep spacing consistent.
I also added in width overrides, which will be useful for form elements.
This allows classes to be added to the HTML to determine the size of the element.