[refs #40] Made recommended changes - code review
I made the recommened changes from the code review.
The app page typography-new was renamed and code in the
typography-old Sass file was moved into typography. The 'old code' in
typography will need to be refactored when we look at components.
For the font-weight, we will look at linting methods.