[refs #107] Add underline when focus on main navigation items
Looking at the main navigation links in high contrast view on Windows,
the yellow outline border is not seen when using the keyboard to
navigate to them and bringing them into focus. To make them standout
and give a visual clue as to where you are, they should be underlined
when in focus.