[refs #186] Accessibility improvements to nav and list elements
Change aria label of nav to "A to Z" rather than "A-Z" so VoiceOver
reads it better.
Add role=list to list so retains semantic meaning as list items are
floated.
Change list to an ordered one. One could argue that the list is an
ordered one as there is a definite and expected order from a to z.
Amend disabled letter from a p element to a span so if CSS is disabled,
shouldn't affect the layout.