[refs #160] Change the way triangle is generated in care cards
Using the clip path method for generating the care card triangles makes
it not show up in High Contrast Mode in Windows. This was picked up in a
DAC report. Despite looking at a gov.uk blog post on this
(https://accessibility.blog.gov.uk/2018/08/01/supporting-users-who-change-colours-on-gov-uk/)
it still didn't seem to work in High Contrast Mode, so it has been
changed to borders and transforms, inspired from
https://codepen.io/mpeace/pres/eBOVGe