[refs #169] Update header Nunjucks macro
Defensively set default 'showNav' and 'showSearch' parameters so if they
are not set, then they default to false. This allows the header macro to
be lazily called with no parameters. So, to just get the header with
logo only, you could get it by simply using:

{{ header({}) }}

Also updated the Nunjucks arguments table to include those not
documented.