Add backstop stage to travis tests
Convert backstop.json configuration to javascript so we can adjust the
test urls for different environments.

Docker for windows and mac supports a "docker.host.internal" address
which points to the host.
Docker for linux doesn't have this feature so the travis CI setup is
slightly more complicated, and must use 127.0.0.1 as the test url.