remove unused styles and markup for conditional reveals
Checkboxes and Radios - remove unused styles and markup for conditional
reveals, as there is no JavaScript code for this functionality these
styles aren't required.

This removes the need for code to check if JavaScript is enabled or
not, with the `js-enabled` class.

Conditional reveals functionality may be added in the future.