Use fallback font if not including NHS font face
There is a configuration option `$nhsuk-include-font-face` which
controls if the NHS corporate font is included in the compiled CSS. In
the current implementation, if this option is false then `font-face` is
not set at all, which leads to unstyled text.

This change makes it so that if the fonts are not included, a
`font-face` declaration is still made using only the fallback fonts (as
set in `$nhsuk-font-fallback`).