[refs #662] Align pagination icon with text
Vertically aligns the pagination icon with the pagination text, even
with large font sizes. This calculation was determined experimentally by
manually positioning the icons at a variety of font sizes, and succeeds
in appearing to vertically align the icon in the majority of cases, with
only very small font sizes looking slightly out of place. The removal of
`margin-top: 0` for `media-type: print` does not appear to affect the
output when printing.