Update footer copyright statement
The copyright guidance was changed in #915 to be copyright of NHS England instead of Crown Copyright, but the text in the footer was only partially updated.

This updates the second instance of Crown Copyright in the footer to also be NHS England.  (The text is in 2 places in the template, and which one is used depends on how many columns of links are included).