Adjust spacing for back link
The back link should be added to the template at the start of the `<div class="nhsuk-width-container">` container but before the `<main>` container. This is to ensure that it gets skipped over when using the "skip to main content" link.

However currently the back link by default will appear too close to the header and too far away from the h1. This adjusts the spacing to add some spacing before the back link, but no spacing after it (as the `<main>` element already add enough vertical spacing at the top using padding).