Move breadcrumb inside nhsuk-width-container
Currently in the examples there’s a `breadcrumb` block which is outside the `<div class="nhsuk-width-container">`, in a similar way to the `beforeContent` block in the template in the NHS design system.

This moves the breadcrumb into `<div class="nhsuk-width-container">`, but before `<main class="nhsuk-main-wrapper" id="maincontent">`, by defining a `outerContent` which matches the existing NHS design system template.