Warning button variant (#976)
On our staff facing service, we use interruption screens that warn a user before they proceed to undertake an action that is destructive and/or cannot be undone. On these pages, we use also a red warning button to further indicate that a destructive action will take place:

<img width="740" alt="Screenshot of a screen asking if a user wants to delete a vaccine." src="https://github.com/nhsuk/nhsuk-frontend/assets/813383/be9fe626-8851-4053-8a7d-925031c8bee2">

This follows [a similar pattern from the GOV.UK Design System](https://design-system.service.gov.uk/components/button/#warning-buttons):

> ### Warning buttons
>
> Warning buttons are designed to make users think carefully before they use them. They only work if used very sparingly. Most services should not need one.

This PR adds an equivalent `--warning` variant to `nhsuk-button`. In the corresponding Service Manual guidance for this component, we can start by using the same guidance as that used in the GOV.UK Design System.