Make it easier to set checked radios and checkboxes
Currently you have to set `checked` on each item.

This would instead make it possible to set `value` (for radios) or `values` (for checkboxes) which would then automatically set the checked values for the items if they match the value given.

The previous method of setting `checked` on each item still works, and would override the default.