Deprecate govuk-grid-column generating class names
The class names generated by the mixin are not flexible enough, as the width is just appended to the class name. In spiking how we might add an extended grid we found the generated class names a little yoda-ish, for example `govuk-grid-column--at-desktop-one-quarter` (bonus points if you actually read that in Yoda’s voice).

Moving the class naming out of the mixin and doing it within the objects layer is also more consistent with the rest of Frontend, and means that users who do use the grid mixins directly don’t have to fit with the constraints of our class name structure.

See: https://github.com/alphagov/govuk-frontend/pull/1090

Co-authored-by: Oliver Byford <oliver.byford@digital.cabinet-office.gov.uk>