Header breaking changes - update nunjucks params (#1109)
This updates the params for the Header nunjucks component to no longer differentiate between 'transactional' and non-transactional services, and instead use the same Nunjucks params for both.

Previously, for the 'non-transactional' service name type, the logo and the service name would be combined as a single link, but for 'transactional' services they would be separate.

To preserve both options, some heuristics are used instead. The service name and logo links are combined in either of these 2 scenarios:

* `logo.href` exactly matches `service.href`
* `service.href` is specified but `logo.href` is unspecified

We’d create some separate guidance on when to combine the logo and service names and when not to, although it will likely be context-dependent and up to teams to decide when to do this based on their research.

This also removes the default link target of `/` for the service name and NHS logo. This allows you to have an option where neither the logo or service name are linked - although this would likely not be recommended.

Finally, there’s also some smaller changes to the nunjucks params to make them more consistent and logical, such as changing `homeHref` to `logo.href`, `service.name` to `service.text` and `organisation.logoURL` to `logo.src`.

## Full list of Nunjucks changes

| Before                      | After        |
|-----------------------------|-------------- |
| `transactionalService.name` | `service.text` |
| `transactionalService.href` | `service.href` |
| `service.name`              | `service.text` |
| `organisation.logoURL`      | `logo.src` |
| `homeHref`                  | `logo.href` |


## Update HTML and CSS classes

The HTML and CSS classes have been updated to more closely follow the Block Element Modifier (BEM) convention.

The previous `nhsuk-header__link` and `nhsuk-header__link--service` have been removed and replaced with:

| Element                        | CSS class                           |
|--------------------------------|-------------------------------------|
| Logo container                 | `.nhsuk-header__logo`               |
| Logo container (if there’s no service name sibling) | `.nhsuk-header__logo--no-service-name`               |
| Logo link                      | `.nhsuk-header__logo__link`         |
| NHS logo itself                | `.nhsuk-logo`                       |
| Service name container         | `.nhsuk-header__service-name`       |
| Service name link              | `.nhsuk-header__service-name__link` |  

---------

Co-authored-by: Colin Rotherham <work@colinr.com>