Add a flask app for testing Jinja rendering
Jinja only allows one template path, so I've symlink things into a combined structure.
This can be more fully automated later on, as this is just a proof of concept.
The same for the template files - ideally this should have a build step like docs.js does.

The flask app should serve the same set of component examples as the javascript
doc site, but render them using Jinja. We can then run all the backstop tests
to check that components rendered with Jinja are visually identical to the
same components rendered with Nunjucks.