Rewrite attributes macro for Jinja compat
There are lots of changes needed to make this Jinja compatable.

Incremental attribute rendering
===============================
Jinja can't modify a variable in a loop and use it outside of the loop, whereas
Nunjucks can't modify attributes of an object inside a loop. The latter means we
cannot use Jinja's namespace object workaround.  To bypass this issue, we now
render the attributes as we calculate them, rather than building them up in an
intermediate variable.

Equality checks
===============
I've weakened the equality checks to == instead of ===, because Jinja doesn't have this operator.
Although this loosens the equality check, I think it's ok here:
1. there is no valid use case for setting `optional` to anything other than a boolean
2. comparing a value to true is a bit stricter than just evaluating the truthiness of the value in an expression, e.g. numbers aren't considered equal to true.

Iterating over a mapping
========================
In Jinja, a mapping is not iterable; you need to call .items() or use the
|items filter.  Nunjucks doesn't have this filter, and the use of .items()
requires nunjucks.installJinjaCompat().

I've chosen not to use installJinjaCompat because it changes lookups to
completely override all calls to `.items`, and this interferes with our use of
`params.items` as a parameter name.  Instead, I've used `dictsort` filter,
which is present in both Jinja and Nunjucks, and returns an iterable.

Concatenation
=============
In Jinja, the + operator must only be used for strings, it does not coerce different types, so ~ is preferable.

If working with escaped strings, and using +, Jinja escapes all operands before
concatenating, but if you use ~ it works as expected in both Jinja and
Nunjucks.

The | safe filter is applied at the end, rather than on the operands to the ~ operator,
because in Nunjucks, concatenating two safe strings creates an unsafe string:

{{'"'|safe}} -> ";
{{'"'|safe + '"'|safe}} -> &quot;

Examples I tested:

    {# works in Jinja only #}
    {{" " ~ name | escape ~ '="' ~ options.value | escape ~ '"'}}
    {{(" " | safe ~ name | escape ~ '="' | safe ~ options.value | escape ~ '"' | safe)}}

    {# works in Nunjucks only #}
    {{(" " + name | escape + '="' + options.value | escape + '"') | safe}}

    {# works in both #}
    {{(" " ~ name | escape ~ '="' ~ options.value | escape ~ '"') | safe}}