Add custom map filter (breaking change)
A map filter in Nunjucks would allow us to write expressions
on arrays of complex objects, rather than looping over the array
and modifying a variable (a pattern which is not supported in Jinja2,
due to its scoping rules).

This commit shows how it could be introduced as a custom filter.

The trade off is that it complicates the usage of the templates, as
extra configuration is needed.