Stop reporting test coverage
Outputting the test code coverage table at the end of each test run makes it harder to spot which tests actually failed, as you have to scroll up each time.

Our test coverage is also pretty bad right now, so the table is full of red, which shows that it’s not really working as a nudge to contributors.

Suggest we remove it for now, work on improving the test coverage separately, and maybe consider a separate tool in future which could check on each pull request whether test coverage has at least not gone down?