Deprecate govuk-grid-row mixin
Wrapping the row functionality up in a mixin offers little other than the ability to use a custom name, but we’re not sure there’s a need for users to do that.

Simplify the code by deprecating the mixin and just using the code to create the concrete class directly.

See: https://github.com/alphagov/govuk-frontend/pull/1090

Co-authored-by: Oliver Byford <oliver.byford@digital.cabinet-office.gov.uk>