Remove fallback implementation of Intl.PluralRules
The intersection of browsers that support ES6 modules (and therefore run our JavaScript) but do not support `Intl.PluralRules` is relatively small:

- Safari 10.3 - 12.5 on iOS
- Safari 10.1 - 12.1 on macOS
- Edge 16-18
- Chrome 61-62
- Opera 48-49

These browsers represent a tiny percentage of users and do not justify the overhead of including the extra JavaScript to implement the fallback logic, so let’s remove it.

These browsers will instead fall back to the ‘other’ translation. For example, when using the Character Count in English it will display ‘You have 1 characters remaining’ rather than the singular translation ‘You have 1 character remaining’.