Add numbered pagination (#1026)
This updates [Pagination component](https://service-manual.nhs.uk/design-system/components/pagination) to support a new variant which is suitable for paging between numbered lists of pages, for example in search results or a long list of patients or appointments.

The variant is based on the [GOV.UK Pagination variant](https://design-system.service.gov.uk/components/pagination/#for-navigating-between-pages-of-items), but with a few changes to match NHS styling, and some differences in the macro params.

The component is backwards-compatible, so the existing previous and next pagination variant is unchanged (both the HTML and Nunjucks options).