Prevent date inputs shifting alignment on iOS 18
Due to an upstream issue in Safari on iOS 18 (and potentially 17), the box model calculations for these inputs go iffy when the value is deleted and re-typed.

The upstream issue was reported during the beta period of iOS 26 and fixed in the release of iOS 26 in September, however no fix for older versions seems to be coming, so we should prevent the issue in our own code in the meantime.

NHS.UK Frontend issue: nhsuk/nhsuk-frontend#1227
Our bug report: alphagov/reported-bugs#90

Co-authored-by: beeps <hi+github@berly.kim>