[v10] Add Interruption panel (#1196)
This adds an Interruption variant of the Panel component, which has a solid blue background and white text.

This is intended for use when you need to stop users in a journey to tell them important information. This should be used vary sparingly, as there is some evidence that users can end up ignoring it, however it could work in some contexts, such as when shown after many regular pages with the off-white background.

---------

Co-authored-by: Colin Rotherham <work@colinr.com>